Last minute refactor on Saturday broke the table test
